Move each extractor into its own module again
